+++ libs/components/lookup/src/lib/modules/autocomplete/autocomplete.component.ts
@@ -363,6 +363,8 @@ export class SkyAutocompleteComponent
       this._inputDirective.blur
         .pipe(takeUntil(this.inputDirectiveUnsubscribe))
         .subscribe(() => {
+          this.inputDirective.restoreInputTextValueToPreviousState();
+          this.closeDropdown();
           this.inputDirective.onTouched();
         });
 
@@ -777,17 +779,6 @@ export class SkyAutocompleteComponent
       this.isOpen = true;
       this.changeDetector.markForCheck();
       this.initOverlayFocusableElements();
-
-      overlay.backdropClick
-        .pipe(takeUntil(this.ngUnsubscribe))
-        .subscribe(() => {
-          /* Sanity check as you should not be able to active the directive from the backdrop */
-          /* istanbul ignore else */
-          if (document.activeElement !== this.inputDirective.inputElement) {
-            this.inputDirective.restoreInputTextValueToPreviousState();
-            this.closeDropdown();
-          }
-        });
     }
   }