+++ libs/components/lookup/src/lib/modules/autocomplete/fixtures/autocomplete.component.fixture.ts
@@ -2,7 +2,7 @@ import { Component, TemplateRef, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 
 import { Subject, of } from 'rxjs';
-import { delay, filter } from 'rxjs/operators';
+import { delay } from 'rxjs/operators';
 
 import { SkyAutocompleteInputDirective } from '../autocomplete-input.directive';
 import { SkyAutocompleteComponent } from '../autocomplete.component';
@@ -19,7 +19,7 @@ import { SkyAutocompleteSelectionChange } from '../types/autocomplete-selection-
 export class SkyAutocompleteFixtureComponent {
   public autocompleteAttribute: string;
 
-  public data: any[] = [
+  public data: { objectid?: string; name?: string; text?: string }[] = [
     { name: 'Red', objectid: 'abc' },
     { name: 'Blue', objectid: 'def' },
     { name: 'Green', objectid: 'ghi' },
@@ -33,7 +33,9 @@ export class SkyAutocompleteFixtureComponent {
     { name: 'Black', objectid: '456' },
   ];
 
-  public model: any = {
+  public model: {
+    favoriteColor: { objectid?: string; name?: string; text?: string };
+  } = {
     favoriteColor: undefined,
   };
   public customNoResultsMessage: string;